requirements.txt
Django~=4.1.4
djangorestframework~=3.14.0
django-ckeditor~=6.6.1
cffi~=1.15.1
Pillow~=9.4.0
pip~=23.2.1
attrs~=22.2.0
pytz~=2022.7.1
PyYAML~=6.0.1
setuptools~=58.1.0
asgiref~=3.6.0
sqlparse~=0.4.3
colorama~=0.4.6
numpy~=1.24.2
requests~=2.28.2
click~=7.1.2
packaging~=23.0
urllib3~=1.26.14
certifi~=2022.12.7
six~=1.16.0
python-dateutil~=2.8.2
inflection~=0.5.1
uritemplate~=4.1.1
idna~=3.4
selenium~=4.8.0
zipp~=3.15.0